requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
openai==1.28.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-core==2.18.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
tinydb==4.8.0
tqdm==4.66.4
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.3
yarl==1.9.4
zipp==3.18.1
Flask-CORS==4.0.1
apscheduler==3.10.4
openai==1.99.6